split ForkBuilder to file.
